feat: Implemented Region's Main Warehouse Command
* Implemented Region's Main Warehouse Command
* Added 'InvalidLocation' Exception
* Refactored Event Handlers
