+++ rushcargo-insiders/src/lib/controller/events.py
@@ -2,16 +2,13 @@ from rich.prompt import Confirm
 import logging
 from rich.logging import RichHandler
 
-from .territoryEvents import territoryEventHandler
-from .buildingEvents import buildingEventHandler
+from .locationEvents import LocationEventHandler
 
 from ..io.arguments import getEventHandlerArguments
-from ..io.constants import (
-    TABLE_TERRITORY_CMD,
-    TABLE_BUILDING_CMD,
-)
+from ..io.constants import TABLE_LOCATION_CMD
 from ..io.validator import clear
 
+from ..model.database import Database
 from ..model.database_tables import console
 
 
@@ -27,20 +24,23 @@ log = logging.getLogger("rich")
 
 # Event Handler Class
 class EventHandler:
+    # Event Handlers
+    __locationEventHandler = None
+
+    # Initialize Event Handler
+    def __init__(self, db: Database, user: str, ORSApiKey: str):
+        # Initialize Location Event Handler
+        self.__locationEventHandler = LocationEventHandler(db, user, ORSApiKey)
+
     # Main Event Handler
     def handler(self, action: str, tableGroup: str, tableName: str) -> None:
         try:
             while True:
                 try:
-                    # Check if it's a Territory Table
-                    if tableGroup == TABLE_TERRITORY_CMD:
-                        # Call Territory Event Handler
-                        territoryEventHandler.handler(action, tableName)
-
-                    # Check if it's a Building Table
-                    elif tableGroup == TABLE_BUILDING_CMD:
-                        # Call Building Event Handler
-                        buildingEventHandler.handler(action, tableName)
+                    # Check if it's a Location Table
+                    if tableGroup == TABLE_LOCATION_CMD:
+                        # Call Location Event Handler
+                        self.__locationEventHandler.handler(action, tableName)
 
                 except Exception as err:
                     console.print(err, style="warning")
