+++ rushcargo-insiders/src/main.py
@@ -8,6 +8,8 @@ from lib.io.constants import *
 from lib.io.arguments import getEventHandlerArguments
 from lib.io.validator import clear
 
+from lib.model.database import initDb
+
 from lib.terminal.constants import TITLE_MSG, WELCOME_MSG
 
 
@@ -38,11 +40,8 @@ def getParserArguments() -> tuple[str, str, str]:
     table = args.table
 
     # Get Table Group
-    if table in TABLE_TERRITORY_CMDS:
-        tableGroup = TABLE_TERRITORY_CMD
-
-    elif table in TABLE_BUILDING_CMDS:
-        tableGroup = TABLE_BUILDING_CMD
+    if table in TABLE_LOCATION_CMDS:
+        tableGroup = TABLE_LOCATION_CMD
 
     return action, tableGroup, table
 
@@ -52,8 +51,11 @@ if __name__ == "__main__":
         # Get argParser Arguments
         action, tableGroup, table = getParserArguments()
 
+        # Initialize Database Connection
+        db, user, ORSApiKey = initDb()
+
         # Initialize Event Handler
-        e = EventHandler()
+        e = EventHandler(db, user, ORSApiKey)
 
         # Call Main Event Handler
         e.handler(action, tableGroup, table)
@@ -68,8 +70,11 @@ if __name__ == "__main__":
             # Get Arguments
             action, tableGroup, table = arguments
 
+            # Initialize Database Connection
+            db, user, ORSApiKey = initDb()
+
             # Initialize Event Handler
-            e = EventHandler()
+            e = EventHandler(db, user, ORSApiKey)
 
             # Call Main Event Handler
             e.handler(action, tableGroup, table)
